+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `ricventu/laravel-anti-xss` will be documented in this file.
+
+## Unreleased
+
+- Initial release: Facade, service singleton, `clean_xss` validation rule and
+  `CleanXss` rule object, opt-in `CleanXssInput` middleware, `@xss` Blade
+  directive and `anti_xss()` helper around [voku/anti-xss](https://github.com/voku/anti-xss).